Differentiable prefix reduction
This commit adds a general differentiable prefix reduction operation.

- ``drjit.prefix_reduce``

See Dr.Jit-Core PR https://github.com/mitsuba-renderer/drjit-core/pull/103
for details on this.

The following two previously existing functions now simply call the general
case.

- ``drjit.prefix_sum``
- ``drjit.cumsum``

There is also a blocked version for 1D arrays:

- ``drjit.blocked_prefix_reduce``

The commit removes sanitization-related options from Dr.Jit, since were
recently moved to Dr.Jit-Core.

+++ src/extra/loop.cpp
@@ -64,7 +64,7 @@ static bool ad_loop_symbolic(JitBackend backend, const char *name,
         // This is necessary to catch cases where a variable is added to the
         // loop state twice.
         read_cb(payload, indices1);
-        for (uint32_t i : indices1)
+        for (uint64_t i : indices1)
             indices2.push_back((uint32_t) i);
         jit_var_loop_update_inner_in(loop.index(), indices2.data());
         indices1.release();
@@ -494,7 +494,7 @@ public:
                 input.is_diff = true;
                 input.has_grad_in = add_index(m_backend, ad_index, true);
                 if (input.has_grad_in)
-                    input.grad_in_index = m_input_indices.size() - 1;
+                    input.grad_in_index = (uint32_t) m_input_indices.size() - 1;
                 input.grad_in_offset = (uint32_t) m_diff_count++;
             }
